updated login test
