+++ src/test/java/edu/ithaca/dturnbull/bank/BankAdmin/BankAdminTest.java
@@ -11,25 +11,6 @@ import edu.ithaca.dturnbull.bank.Teller.BankTeller;
 
 public class BankAdminTest {
 
-    @Test
-    void loginTest(){
-        Bank bank = new Bank();
-
-        Customer customer = new Customer(0, "password");
-        bank.addCustomer(customer);
-        Customer customerTest = bank.customerLogIn(0, "password");
-        assertEquals(customer, customerTest);
-
-        AbstractTeller teller = new BankTeller(0, "password");
-        bank.addTeller(teller);
-        AbstractTeller tellerTest = bank.tellerLogIn(0, "password");
-        assertEquals(teller, tellerTest);
-
-        BankAdmin admin = new BankAdmin(0, "password");
-        bank.addAdmin(admin);
-        BankAdmin adminTest = bank.adminLogIn(0, "password");
-        assertEquals(admin, adminTest);
-
-    }
+    
     
 }
